Move ADTypeCheckContext tests to a separate module
