+++ test/runtests.jl
@@ -27,6 +27,10 @@ macro timeit_include(path::AbstractString)
 end
 
 @testset "Turing" begin
+    @testset "Test utils" begin
+        @timeit_include("test_utils/test_utils.jl")
+    end
+
     @testset "Aqua" begin
         @timeit_include("Aqua.jl")
     end
