+++ test/test_utils/ad_utils.jl
@@ -229,44 +229,6 @@ function DynamicPPL.dot_tilde_observe(context::ADTypeCheckContext, sampler, righ
     return logp, vi
 end
 
-# Check that the ADTypeCheckContext works as expected.
-Test.@testset "ADTypeCheckContext" begin
-    Turing.@model test_model() = x ~ Turing.Normal(0, 1)
-    tm = test_model()
-    adtypes = (
-        Turing.AutoForwardDiff(),
-        Turing.AutoReverseDiff(),
-        Turing.AutoZygote(),
-        # TODO: Mooncake
-        # Turing.AutoMooncake(config=nothing),
-    )
-    for actual_adtype in adtypes
-        sampler = Turing.HMC(0.1, 5; adtype=actual_adtype)
-        for expected_adtype in adtypes
-            if (
-                actual_adtype == Turing.AutoForwardDiff() &&
-                expected_adtype == Turing.AutoZygote()
-            )
-                # TODO(mhauru) We are currently unable to check this case.
-                continue
-            end
-            contextualised_tm = DynamicPPL.contextualize(
-                tm, ADTypeCheckContext(expected_adtype, tm.context)
-            )
-            Test.@testset "Expected: $expected_adtype, Actual: $actual_adtype" begin
-                if actual_adtype == expected_adtype
-                    # Check that this does not throw an error.
-                    Turing.sample(contextualised_tm, sampler, 2)
-                else
-                    Test.@test_throws AbstractWrongADBackendError Turing.sample(
-                        contextualised_tm, sampler, 2
-                    )
-                end
-            end
-        end
-    end
-end
-
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # List of AD backends to test.
